Fix the two type-gate defects PR #27 merged with still failing
PR #27 (merged as e9ecbab) landed with its own gate red: 12 hard-layer
errors and a ratchet that undercounted the soft layer. Both are fixed here,
using the numbers from that PR's own CI run
(https://github.com/ilegault/RBL/actions/runs/34187056116, jobs
test (3.13) and test (3.14), 2026-09-08 04:27-04:32Z, windows-latest):

- 10 of the 12 hard-layer errors are genuine stub gaps: pyvisa's `Resource`
  and cv2's module surface don't declare the attributes funcgen_driver.py
  (9 errors) and camera_source.py (1, `VideoWriter_fourcc`) use, so no
  annotation in this repo can fix them. A `[[tool.mypy.overrides]]` block
  scoped to exactly those two modules disables only `attr-defined`; every
  other error class in either module is still reported and still fails the
  build (the exclude list this replaced governed directory discovery, not
  imports, and is not being reintroduced under a new name).
- The other 2 were real bugs, fixed directly rather than suppressed:
  load_model.py:135 multiplied an unconverted np.trapezoid result (typed
  _ArrayOrScalarCommon) by a float; wrapping it in float() first resolves
  the operand mismatch with no behavior change (verified: all 21
  test_load_model.py cases still pass). camera_source.py's _latest_frame
  was assigned `None` with no annotation, so mypy inferred it as
  `None`-only and rejected every real frame written to it — annotated as
  `object | None` to match how this class already types its Qt signals.
- tools/mypy_ratchet.txt said 132; that PR's CI run measured 140 soft-layer
  errors. Corrected to 140 from that run's own log, not a local one — a
  local `mypy` run without this repo's pinned dependencies installed
  undercounts (missing packages resolve to Any instead of triggering real
  errors), which is how 132 got committed in the first place.

Verified locally with the full requirements-dev.txt installed: `mypy`
reports exactly 140 errors with 0 under rbl/config or rbl/hardware, so
`tools/type_gate.py` exits 0 (0 hard, 140 soft against a ratchet of 140).
ruff still passes.

Also brings in baseline.md (pushed directly to master as commit 2ba4b6b
after PR #27 merged, recording that run's test/type/lint results) and
updates ticket 01 to Status: done with its checklist and a Comments entry
noting the baseline was written after the fact from PR #27's own logs.

+++ src/rbl/hardware/load_model.py
@@ -130,7 +130,7 @@ def capacitance_from_charge(current_ma, dt_s: float, baseline_ma: float,
     x = np.asarray(current_ma, dtype=float) - baseline_ma
     if x.size < 2 or delta_v_kv == 0:
         return float("nan")
-    q_coulombs = np.trapezoid(x * 1e-3, dx=dt_s)
+    q_coulombs = float(np.trapezoid(x * 1e-3, dx=dt_s))
     delta_v_v = abs(delta_v_kv) * 1000.0
     return abs(q_coulombs / delta_v_v) * 1e12
